--- .dockerignore
@@ -1,5 +1,0 @@
-dev
-.gitignore
-LICENSE
-README.md
-dist/main.db
