+++ .env.example
@@ -9,15 +9,3 @@ DOMAIN_BASE=http://localhost:8080
 CACHE_MAP_LIMIT=15
 # Maximum API call requests to the API, this will ban the api for 5 minutes after requesting more than API_RATE_LIMIT in 5 minutes.
 API_BAN_LIMIT=300
-
-# Optional database driver, you can just ignore and use sqlite
-# ! Warning: Sqlite not implemented yet
-DATABASE_DRIVER=mariadb
-
-# Mysql database connetion details, fill these in if you chose mysql above.
-MYSQL_ROOT_PASSWORD=example-dev
-MYSQL_DATABASE=freepad
-MYSQL_USER=freepad
-MYSQL_PASSWORD=example-dev
-MYSQL_URL=mariadb
-MYSQL_PORT=3306
