--- docker-compose.example.yaml
@@ -1,38 +1,0 @@
-version: '3.4'
-services:
-
-  freepad:
-    build: .
-    networks:
-      - backend
-    environment:
-      - DOMAIN_BASE=http://localhost:8888
-      - MYSQL_URL=mariadb
-      - MYSQL_ROOT_PASSWORD
-      - MYSQL_DATABASE
-      - MYSQL_USER
-      - MYSQL_PASSWORD
-      - MYSQL_PORT
-    depends_on:
-      - mariadb
-    ports:
-      - 8888:8080
-
-  mariadb:
-    image: mariadb:10.2
-    environment:
-      - MYSQL_ROOT_PASSWORD
-      - MYSQL_DATABASE
-      - MYSQL_USER
-      - MYSQL_PASSWORD
-    restart: unless-stopped
-    networks:
-      - backend
-    volumes:
-     - ./dev/mariadb:/var/lib/mysql
-    ports:
-      - 3306:3306
-
-networks:
-  backend:
-    driver: bridge
