--- src/controllers/post/post_types.go
@@ -1,6 +1,0 @@
-package post
-
-type Post struct {
-	Name    string `json:"name"`
-	Content string `json:"content"`
-}
