--- src/types/types_error.go
@@ -1,6 +1,0 @@
-package types
-
-type FreeError struct {
-	Error   string `json:"error"`
-	Message string `json:"message"`
-}
